merge

+++ test/Microsoft.ML.Tests/Transformers/ValueMappingTests.cs
@@ -559,7 +559,7 @@ namespace Microsoft.ML.Tests.Transformers
                 Assert.True(result.Schema.TryGetColumnIndex("GroupId", out int groupIdx));
 
                 Assert.True(result.Schema[labelIdx].Type is KeyType);
-                Assert.Equal(5, result.Schema[labelIdx].Type.GetItemType().GetKeyCount());
+                Assert.Equal((ulong)5, result.Schema[labelIdx].Type.GetItemType().GetKeyCount());
 
                 var t = result.GetColumn<uint>(Env, "Label");
                 uint s = t.First();
